changed main

+++ models.py
@@ -40,7 +40,7 @@ class Sale(Base):
     __tablename__ = "sale"
 
     id = sq.Column(sq.Integer, primary_key=True)
-    price = sq.Column(sq.Integer, nullable=False)
+    price = sq.Column(sq.Float(2), nullable=False)
     date_sale = sq.Column(sq.Date, nullable=False)
     id_stock = sq.Column(sq.Integer, sq.ForeignKey("stock.id"), nullable=False)
     count = sq.Column(sq.Integer, nullable=False)
